Eliminated useless TestIdentityConfiguration class.
Moved the static methods to TestUserIdentities and deleted the empty shell.
