+++ backend/src/test/java/gov/cdc/usds/simplereport/test_util/TestUserIdentities.java
@@ -1,6 +1,11 @@
 package gov.cdc.usds.simplereport.test_util;
 
 import gov.cdc.usds.simplereport.service.model.IdentityAttributes;
+import java.util.List;
+import org.springframework.security.authentication.TestingAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 public class TestUserIdentities {
 
@@ -11,4 +16,29 @@ public class TestUserIdentities {
       new IdentityAttributes(STANDARD_USER, "Bobbity", "Bob", "Bobberoo", null);
   public static final IdentityAttributes SITE_ADMIN_USER_ATTRIBUTES =
       new IdentityAttributes(SITE_ADMIN_USER, "Ruby", "Raven", "Reynolds", null);
+  /**
+   * Set the security context to hold a particular user, then run some code, then reset the user.
+   *
+   * @param username a username that is known to the current IdentitySupplier
+   * @param nested a lambda to run
+   */
+  public static void withUser(String username, Runnable nested) {
+    SecurityContext context = SecurityContextHolder.getContext();
+    Authentication original = context.getAuthentication();
+    try {
+      context.setAuthentication(new TestingAuthenticationToken(username, null, List.of()));
+      nested.run();
+    } finally {
+      context.setAuthentication(original);
+    }
+  }
+  /**
+   * Set the security context to hold the "default" user ({@link STANDARD_USER}),
+   * then run some code, then reset the user.
+   *
+   * @param nested a lambda to run
+   */
+  public static void withStandardUser(Runnable nested) {
+    withUser(STANDARD_USER, nested);
+  }
 }
